chore: add link converter

+++ backend/src/main/java/io/github/xpakx/outline/repo/LinkRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LinkRepository extends JpaRepository<Link, Integer> {
+public interface LinkRepository extends JpaRepository<Link, Long> {
 }
